Add Uniswap V3 / G-UNI strategy

+++ contracts/strategies/liquity/interfaces/IPriceFeed.sol
@@ -1,8 +1,8 @@
-// SPDX-License-Identifier: MIT
+// SPDX-License-Identifier: GPL-3.0-or-later
 
 pragma solidity 0.8.6;
 
 interface IPriceFeed {
     // --- Function ---
-    function fetchPrice() external returns (uint);
+    function fetchPrice() external returns (uint256);
 }
